app/requirements.txt
asgiref==3.8.1
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
iniconfig==2.0.0
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg==3.2.4
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
sqlparse==0.5.3
typing_extensions==4.12.2
